bot/requirements.txt
discord.py>=2.4.0
git+https://github.com/imayhaveborkedit/discord-ext-voice-recv.git
aiohttp
numpy
scipy
pynacl
redis